% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Calc_MSYAndBiolRefPoints}
\alias{Calc_MSYAndBiolRefPoints}
\title{Calculate MSY and BMSY}
\usage{
Calc_MSYAndBiolRefPoints(mod_type, mod_option, param_results, env)
}
\arguments{
\item{mod_type}{1=Schaefer, 2=Fox, 3=Pella_Tomlinson}

\item{mod_option}{1=traditional, 2=chl, 3=dep, 4=chl and dep}

\item{param_results}{object containing results outputted by for model by TMB}

\item{env}{standardised environmental index value. 0=average conditions.}
}
\value{
Bmsy1, MSY1, F_lim, F_thresh, F_targ, B_lim, B_thresh, B_targ
}
\description{
This function calculates maximum sustainable yield (MSY) and biomass at MSY (BMSY),
for each of the model types and model options
}
