% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Calc_variance_covariance_matrix}
\alias{Calc_variance_covariance_matrix}
\title{Get variance-covariance matrix, for use in resampling}
\usage{
Calc_variance_covariance_matrix(obj_TMB)
}
\arguments{
\item{obj_TMB}{outputs outputted directly from fitted TMB biomass dynamics model}

\item{DatFromCSVFile}{model data read in from csv file}

\item{bdm_params}{object containing parameters for TMB model}

\item{model_outputs}{key model outputs and associated asymptotic variances}

\item{nreps}{specified number of random values for each parameter produced by parametric resampling}

\item{use60CL_for_Biom}{flag to output 95 or 60 percent confidence limits. False = 95 percent, True = 60 percent}
}
\value{
Params, vcov.Params
}
\description{
This function outputs variance-covariance matrix, dervied from Hessian matrix outputted by TMB,
for use in parametric resampling of model parameters to generate estimates of uncertainty, i.e. without
use of asymptotic error approximation. The Params and vcov.Params objects can be passed to the
function Get_model_outputs_from_resampling()
}
