% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{fit_the_model}
\alias{fit_the_model}
\title{Fit the biomass dynamics model using TMB}
\usage{
fit_the_model(
  DatFromCSVFile,
  mod_scenario,
  mod_type,
  mod_option,
  bdm_data,
  bdm_params,
  fix_pInit,
  bdm_map = NULL,
  initial_params = NULL,
  nyrs
)
}
\arguments{
\item{DatFromCSVFile}{Data object read in from a csv file, containing all model time series data}

\item{mod_scenario}{1=non-state space, 2=state space}

\item{mod_type}{1=Schaefer, 2=Fox, 3=Pella_Tomlinson}

\item{mod_option}{1=traditional, 2=chl, 3=dep, 4=chl and dep}

\item{bdm_data}{object containing input data for TMB model}

\item{bdm_params}{object containing parameters for TMB model}

\item{fix_pInit}{map option to fix pInit parameter 1=yes, 0=no}

\item{bdm_map}{object containing parameter map for TMB model}

\item{initial_params}{optional input to input starting values when fitting state space model}
}
\value{
multiple objects with output from TMB  (rep_TMB, fit_TMB, obj_TMB, best_par_TMB, best_fitted_par_TMB, best_val_TMB, new_bdm_map)
}
\description{
The is a wrapper function used to fit the biomass dynamics model using TMB.
The model is always first fitted as a non-state space model, to get better
starting values of key parameters before fitting the state space model.
}
