% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_BDM_Data}
\alias{Get_BDM_Data}
\title{Get input data for the biomass dynamics model}
\usage{
Get_BDM_Data(
  DatFromCSVFile,
  wt_param_pen,
  wt_depl_pen,
  wt_biom_pen,
  wt_harv_pen,
  wt_cpue1,
  wt_cpue2,
  wt_cpue3,
  wt_cpue4,
  SigmaR,
  Sigma_env,
  max_currBrel
)
}
\arguments{
\item{DatFromCSVFile}{Data object read in from a csv file, containing all model time series data}

\item{wt_param_pen}{lambda for parameter penalty}

\item{wt_depl_pen}{lambda for final depletion penalty}

\item{wt_biom_pen}{lambda for biomass penalty}

\item{wt_harv_pen}{lambda for harvest rate penalty}

\item{wt_cpue1}{lambda for obs_ln_cpue1 objective function component}

\item{wt_cpue2}{lambda for obs_ln_cpue2 objective function component}

\item{wt_cpue3}{lambda for obs_ln_cpue3 objective function component}

\item{wt_cpue4}{lambda for obs_ln_cpue4 objective function component}

\item{SigmaR}{specified level of process error associated with biomass production}

\item{Sigma_env}{specified level of process error associated with environment}

\item{max_currBrel}{maximum level of relative biomass for current year, used in penalty for final depletion level}
}
\value{
object containing all data required as model input
}
\description{
This function produces a list containing all of the data objects required as input for the model
}
