% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_BDM_Map}
\alias{Get_BDM_Map}
\title{Get model template model builder (TMB) parameter map}
\usage{
Get_BDM_Map(DatFromCSVFile, mod_scenario, mod_option)
}
\arguments{
\item{DatFromCSVFile}{Data object read in from a csv file, containing all model time series data}

\item{mod_scenario}{1=non-state space, 2=state space}

\item{mod_option}{1=traditional, 2= +chl, 3= +dep, 3 = +chl and dep}
}
\value{
List of model parameters that should not be estimated using TMB
}
\description{
This function produces an list, listing model parameters that should not be estimated using TMB. Note
that for the second scenario, EpsR are 'estimated' as random effects.
}
\examples{
mod_scenario = 2 # 1=non-state space, 2=state space
model_option = 1 # 1=traditional, 2= +chl, 3= +dep, 3 = +chl and dep
nyrs = 44
Get_BDM_Map(DatFromCSVFile, mod_option, mod_scenario)
}
