% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_model_outputs}
\alias{Get_model_outputs}
\title{This function returns a range of outputs for the fitted model}
\usage{
Get_model_outputs(result, nyrs)
}
\arguments{
\item{result}{results of fitted TMB model}

\item{nyrs}{number of years of data}
}
\value{
model_outputs, object containing TMB outputs for plotting
}
\description{
This function returns a range of outputs for the fitted model (required for plotting)
}
