% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_model_outputs_from_resampling}
\alias{Get_model_outputs_from_resampling}
\title{Output model results from parametric resampling}
\usage{
Get_model_outputs_from_resampling(
  DatFromCSVFile,
  bdm_params,
  obj_TMB,
  VarCov_res,
  model_outputs,
  nreps,
  use60CL_for_Biom
)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{bdm_params}{object containing parameters for TMB model}

\item{obj_TMB}{outputs outputted directly from fitted TMB biomass dynamics model}

\item{VarCov_res}{estimated params and variance covariance matrix, from inverted Hessian matrix}

\item{model_outputs}{key model outputs and associated asymptotic variances}

\item{nreps}{specified number of random values for each parameter produced by parametric resampling}

\item{use60CL_for_Biom}{flag to output 95 or 60 percent confidence limits. False = 95 percent, True = 60 percent}
}
\value{
plot of estimated random effects
}
\description{
This function outputs model results derived by parametric resampling, i.e. by passing
random values of model parameters, from the estimated variance-covariance matrix, through
the model and retrieving results, assuming the estimated parameters conform to a multivariate normal distribution
}
