% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_model_parameter_outputs}
\alias{Get_model_parameter_outputs}
\title{This function returns a range of model parameter outputs for the fitted model}
\usage{
Get_model_parameter_outputs(result, mod_type, mod_option)
}
\arguments{
\item{result}{results of fitted TMB model}

\item{mod_type}{1=Schaefer, 2=Fox, 3=Pella_Tomlinson}

\item{mod_option}{1=traditional, 2=chl, 3=dep, 4=chl and dep}
}
\value{
outputs, object containing TMB outputs for plotting
}
\description{
This function returns a range of model parameter outputs for the fitted model
}
