% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Get_MSY_and_BMSY_with_err}
\alias{Get_MSY_and_BMSY_with_err}
\title{Output  MSY and BMSY estimates from parametric resampling}
\usage{
Get_MSY_and_BMSY_with_err(mod_type, mod_option, ResampRes, env)
}
\arguments{
\item{mod_type}{1=Schaefer, 2=Fox, 3=Pella_Tomlinson}

\item{mod_option}{1=traditional, 2=chl, 3=dep, 4=chl and dep}

\item{ResampRes}{object containing results outputted by for model by TMB, after resampling}

\item{env}{standardised environmental index value. 0=average conditions.}
}
\value{
MSY_sim_res, estimates of MSY and BMSY, including mean, median, lower and upper 60 and 95 percentiles
}
\description{
This function outputs MSY and BMSY estimates from parametric resampling, using model parameter values
derived by passing  random values of model parameters, from the estimated variance-covariance matrix, through
MSY and BMSY calculutions, assuming the estimated parameters conform to a multivariate normal distribution
}
