% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{GetSDrep}
\alias{GetSDrep}
\title{Get values calculated by ADREPORT statements in the TMB C++ code}
\usage{
GetSDrep(obj)
}
\arguments{
\item{obj}{object containing TMB output}
}
\value{
object containing TMB output
}
\description{
This function sets up the TMB function to avoid NLL from fn() differing from
report and failing to match calculated value with penalties. The values calculated by ADREPORT
statements in the TMB C++ code are returned. These may be erroneous if the model has not converged.
Note, obj = obj_TMB.
}
