% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Biomass_Catch_And_Exploitation}
\alias{Plot_Biomass_Catch_And_Exploitation}
\title{Plot biomass, catches and exploitation (asymptotic error)}
\usage{
Plot_Biomass_Catch_And_Exploitation(
  DatFromCSVFile,
  mod_type,
  mod_option,
  model_outputs,
  param_results,
  xaxis_lab
)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{mod_type}{1=Schaefer, 2=Fox, 3=Pella_Tomlinson}

\item{mod_option}{1=traditional, 2=chl, 3=dep, 4=chl and dep}

\item{model_outputs}{object containing outputs from fitted TMB model}

\item{param_results}{object containing outputs from fitted TMB model}

\item{xaxis_lab}{x axis label, to overide default}
}
\value{
plots of biomass, catch and exploitation
}
\description{
This function plots relative and absolute biomass, catch and exploitation
and associated 95 percent confidence limits (derived from asymptotic errors)
}
