% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Estimated_Random_Effects}
\alias{Plot_Estimated_Random_Effects}
\title{Plot estimated random effects}
\usage{
Plot_Estimated_Random_Effects(DatFromCSVFile, model_outputs)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{model_outputs}{object containing outputs from fitted TMB model}
}
\value{
plot of estimated random effects
}
\description{
This function plots estimated random effects from the state space production model
}
