% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Kobe_Plot}
\alias{Plot_Kobe_Plot}
\title{Plot exploitation vs relative biomass (i.e. Kobe plot)}
\usage{
Plot_Kobe_Plot(DatFromCSVFile, model_outputs)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{model_outputs}{object containing outputs from fitted TMB model}
}
\value{
Kobe plot
}
\description{
This function plots exploitation vs relative biomass (i.e. Kobe plot)
}
