% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Mod_Car_Capacity}
\alias{Plot_Mod_Car_Capacity}
\title{Plot derived values for annual carrying capacity (if assuming environment directly affects K)}
\usage{
Plot_Mod_Car_Capacity(DatFromCSVFile, result_TMB, xaxis_lab = NA, y_max = NA)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{result_TMB}{object containing outputs from fitted TMB model}

\item{xaxis_lab}{label for x axis}

\item{y_max}{max y axis value}
}
\value{
plot of dervied annual K values
}
\description{
This plots derived values for annual carrying capacity, using outputs from the fitted
state space production model.
}
