% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Obs_vs_Exp_CPUE_resamp}
\alias{Plot_Obs_vs_Exp_CPUE_resamp}
\title{Plot observed vs expected CPUE (parametric resampling)}
\usage{
Plot_Obs_vs_Exp_CPUE_resamp(DatFromCSVFile, ResampRes, xaxis_lab = "Year")
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{ResampRes}{object containing outputs from parametric resampling}

\item{xaxis_lab}{x axis label, to overide default}
}
\value{
plots of observed vs expected cpue
}
\description{
This function plots observed vs expected CPUE and associated 95 percent confidence limits
based on parametric resampling
}
