% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Plot_Obs_vs_Exp_Env_Index}
\alias{Plot_Obs_vs_Exp_Env_Index}
\title{Plot observed vs expected environmental index}
\usage{
Plot_Obs_vs_Exp_Env_Index(
  DatFromCSVFile,
  model_outputs,
  xaxis_lab = NA,
  y_max = NA,
  y_min = NA
)
}
\arguments{
\item{DatFromCSVFile}{model data read in from csv file}

\item{model_outputs}{object containing outputs from fitted TMB model}

\item{xaxis_lab}{x axis label, to overide default}

\item{y_max}{y axis maximum, to overide default}

\item{y_min}{y axis minimum, to overide default}
}
\value{
plots of observed vs expected environmental index
}
\description{
This function plots observed vs expected environmental index and associated 95 percent confidence limits
}
