% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Set_BoundsForBDM_Params}
\alias{Set_BoundsForBDM_Params}
\title{Set lower and upper bounds for estimated model parameters}
\usage{
Set_BoundsForBDM_Params(
  ln_K_bnds,
  ln_r_bnds,
  ln_q_bnds,
  ln_sd_bnds,
  env_param_bnds,
  ln_dep_bnds,
  ln_pt_bnds,
  pInit_bnds
)
}
\arguments{
\item{ln_K_bnds}{natural log of initial value for population carrying capacity.}

\item{ln_r_bnds}{natural log of initial value for population intrinsic increase.}

\item{env_param_bnds}{natural log of initial value for catchability param for 3rd CPUE series. Set to NA if 2 CPUE series.}

\item{pInit_bnds}{initial level of relative biomass for population.}

\item{ln_q_params_bnds}{natural log of initial value for catchability param for 1st CPUE series.}

\item{ln_sdmod_params_bnds}{natural log of initial value for catchability param for 2nd CPUE series. Set to NA if only 1 CPUE series.}

\item{ln_dep_param_bnds}{natural log of initial value for catchability param for 4th CPUE series. Set to NA if 3 CPUE series.}
}
\value{
object containing lower and upper bounds of model parameters for input into the model
}
\description{
This function produces a lists for lower and upper bounds estimated model parameters.
Values are specified for all parameters, whether estimated or not.
}
\examples{
ln_K_bnds <- c(6,9)
ln_r_bnds <- log(c(0.01,1.5))
ln_q_bnds <- c(-20,0)
ln_sd_bnds <- c(-20,1)
env_param_bnds <- c(-20,20)
ln_dep_bnds <- c(-20,0)
ln_pt_bnds <- c(-20,0)
pInit_bnds = c(0,1)
bdm_param_bounds = Set_BoundsForBDM_Params(ln_K_bnds, ln_r_bnds, ln_q_bnds, ln_sd_bnds,
                                         env_param_bnds, ln_dep_bnds, ln_pt_bnds, pInit_bnds)
}
