% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{Set_InitValsForBDM_Params}
\alias{Set_InitValsForBDM_Params}
\title{Set initial values for estimated model parameters}
\usage{
Set_InitValsForBDM_Params(
  ln_K,
  ln_r,
  ln_q1,
  ln_q2,
  ln_q3,
  ln_q4,
  ln_sd1,
  ln_sd2,
  ln_sd3,
  ln_sd4,
  env_p,
  lndep,
  ln_pt_parm,
  FF,
  EpsR
)
}
\arguments{
\item{ln_K}{natural log of initial value for population carrying capacity.}

\item{ln_r}{natural log of initial value for population intrinsic increase.}

\item{ln_q1}{natural log of initial value for catchability param for 1st CPUE series.}

\item{ln_q2}{natural log of initial value for catchability param for 2nd CPUE series.}

\item{ln_q3}{natural log of initial value for catchability param for 3rd CPUE series.}

\item{ln_q4}{natural log of initial value for catchability param for 4th CPUE series.}

\item{lndep}{depensation parameter. In normal space, 0=no depensation, 1=full depensation.}

\item{FF}{exploitation parameters.}

\item{EpsR}{random effects parameters.}

\item{ln_sdmod1}{natural log of initial value for sd for 1st CPUE series.}

\item{ln_sdmod2}{natural log of initial value for sd for 2nd CPUE series.}

\item{ln_sdmod3}{natural log of initial value for sd for 3rd CPUE series.}

\item{ln_sdmod4}{initial value for sd for 4th CPUE series. Set to zero for no direct effect of environmental parameter.}

\item{env_param}{natural log of initial value for environmental param linked to biomass.}

\item{ln_SigmaR}{level of process error.}
}
\value{
object containing initial values of model parameters for input into the model
}
\description{
This function produces a list of initial values for estimated model parameters.
The parameters listed depends on number of CPUE series. Note, when fitting models
without any link to an environmental index, the value should be set to zero. Set
initial values for all parameters, whether estimated or not.
}
\examples{
nyrs = 44
ln_K = log(8000)
ln_r = log(0.5)
ln_q1 = -10
ln_q2 = -10
ln_q3 = -10
ln_q4 = -10
ln_sdmod1 = NA
ln_sdmod2 = NA
ln_sdmod3 = NA
ln_sdmod4 = NA
env_param = 0
lndep = -1
FF = rep(-2, nyrs)
EpsR = rep(0, times=nyrs)
ln_SigmaR = log(0.3)
bdm_params = Set_InitValsForBDM_Params(ln_K, ln_r, ln_q1, ln_q2, ln_q3, ln_q4, ln_sd1, ln_sd2,
                                       ln_sd3,ln_sd4, env_p, lndep, ln_pt_parm, FF, EpsR)
}
