% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_BDM.R
\name{update_full_parameter_set}
\alias{update_full_parameter_set}
\title{Update the full set of parameters for the model with the values estimated in TMB}
\usage{
update_full_parameter_set(full_parameter_set, best_param_estimates)
}
\arguments{
\item{obj}{object containing TMB output}
}
\value{
object containing TMB output
}
\description{
This function updates the full set of parameters for the model with the values estimated in TMB,
from the TMB object obj$env$last.par.best
}
\keyword{internal}
